Add messages to the app
Added messages to the base template. Multiple messages can be displayed at the same time because a for loop was used for generating messages. There are 3 types of messages whose appearances were customized: error, success, and info. Each message type has a different border color, but other stylistic aspects are the same. Added a message to display if the city was not found.