Added option to add country for weather searching
Changed the way the view processes data. Now, when the user provides a city name followed by a country name, the app will recognize the data and include the country in the search for weather conditions. This allows for a more precise search because one city could exist in multiple countries (e.g., London in the UK and London in the USA). The app uses the pycountry library to get the ISO code of the country, which is then passed to the API request.